refactor(election-map): remove helper folder inside component folder
